refactor(entity/settings): move theme presets to shared layer
